% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigsustainode.R
\name{pigsustainode}
\alias{pigsustainode}
\title{A function to fit \pkg{pigsustainr} ODE models}
\usage{
pigsustainode(model_name, initial_values, times, backend)
}
\arguments{
\item{model_name}{The name of the model to simulate}

\item{initial_values}{Initial values of the state variables}

\item{times}{The simulation time steps. Should increment by the desired
dt interval}

\item{backend}{Either "deSolve" or "cpp" (see details)}
}
\value{
An object of class of \code{PigSustainSim}
}
\description{
Fit \pkg{pigsustainr} ODE (ordinary differential equation) models, using
either the 'deSolve' R package or the C++.
}
\author{
Conor Goold \email{conor.goold@gmail.com}
}
