\name{pigsustainr-package}
\alias{pigsustainr-package}
\alias{pigsustainr}
\docType{package}
\title{\packageTitle{pigsustainr}}
\description{\packageDescription{pigsustainr}}
\section{Package Content}{\packageIndices{pigsustainr}}
\author{\packageAuthor{pigsustainr}}
\section{Maintainer}{\packageMaintainer{pigsustainr}}
\keyword{package}
