% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PigSustainSens-class.R
\docType{class}
\name{PigSustainSens-class}
\alias{PigSustainSens-class}
\alias{PigSustainSens}
\title{An S4 class to represent \pkg{pigsustainr} sensitivity analysis results}
\description{
Sensitivity analyses of ODE models inspect how the relative change in
the ODE solutions to small changes in the parameter values. Formally,
one method of sensitivity analysis is to compute the partial derivative
of the solution with respect to the parameters.
}
