% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigsustainsens.R
\name{pigsustainsens}
\alias{pigsustainsens}
\title{A function to perform sensitivity analyses on \pkg{pigsustainr} ODE models}
\usage{
pigsustainsens(
  model_name,
  initial_values,
  base_parameters,
  times,
  backend = "cpp",
  sensitivity_parameters = NULL,
  perturb_prop = 0.01,
  n_sims = 1
)
}
\description{
Local sensitivity is defined as the partial derivative of the state variables with respect
to a change in a parameter. Although this is analytically tractable in many cases,
we can approximate for any ODE model computationally, which is what the function
below provides. The raw sensitivity value is multiplied by
}
