% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PigSustainSim-class.R
\docType{class}
\name{PigSustainSim-class}
\alias{PigSustainSim-class}
\alias{PigSustainSim}
\title{An S4 class for simulated \pkg{pigsustainr} models}
\description{
Simulation models in \pkg{pigsustainr} are represented as instances of
\code{PigSustainSim}, which hold the model type that was run,
the results of the simulations by time, the parameters used to run the model,
and any relevant diagnostic information.
}
\section{Slots}{

\describe{
\item{\code{model}}{The name of the model to use}

\item{\code{results}}{A matrix of simulation results}

\item{\code{parameters}}{A vector containing the parameters used to run the model}
}}

