% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PigSustainSim-class.R
\name{summary,PigSustainSim-method}
\alias{summary,PigSustainSim-method}
\title{Retrieve the simulation results from a
PigSustainSim class instance.}
\usage{
\S4method{summary}{PigSustainSim}(object)
}
\description{
Retrieve the simulation results from a
PigSustainSim class instance.
}
