% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukpork}
\alias{ukpork}
\title{UK pork industry data at monthly resolution between 2015 and 2019}
\format{
A dataframe with 60 values and 15 columns
}
\usage{
data(ukpork)
}
\description{
The dataset contains monthly values for the number of sows, sows in-pig, 
piglets, weaners, growers and finishers, as well as total new monthly supply, 
imports and exports, and pig prices.
}
\keyword{datasets}
