% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integrate_ode}
\alias{integrate_ode}
\title{Integrates a PigSustainR model}
\usage{
integrate_ode(model_type, times, inits, parameters, dt)
}
\arguments{
\item{model_type}{A string for the model type to integrate}

\item{times}{A vector of integration time steps}

\item{inits}{A vector of initial values of the state variables}

\item{parameters}{A vector of parameters}
}
\value{
A dataframe of state variable solutions
}
\description{
Integrates a PigSustainR model
}
