% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigsustainr-package.R
\docType{package}
\name{pigsustainr-package}
\alias{pigsustainr-package}
\alias{pigsustainr}
\title{Dynamic simulation models for investigating the sustainability of pig industries}
\description{
The \pkg{pigsustainr} package provides a number of dynamic
system simulation models to model the sustainability of pig industries.
}
